+++ src/main.rs
@@ -9,7 +9,6 @@ mod hash;
 mod hashfile;
 mod herder_api;
 mod herder_daemon;
-mod herder_facade;
 mod ipc_common;
 mod logging;
 mod native;
