+++ src/orchestrator/herder_facade/client.rs
@@ -1,5 +1,5 @@
+use super::DaemonError;
 use crate::herder_api::StartHerd;
-use crate::herder_facade::DaemonError;
 use crate::ipc_common::write_msg_async;
 use serde::Serialize;
 use tokio::io::AsyncWrite;
